// src/Enums/Enums.ts
export enum PagesEnum {
    Home = 'Home',
    About = 'About',
    News = 'News',
    Payment = 'Payment',
    FAQ = 'FAQ',
    Charity = "Charity",
    Auth = "Auth",
    Admin = "Admin"
}

export enum FormsEnum {
    AskQuestion = "AskQuestion",
    Donate = "Donate",
    CarNumber = "CarNumber",
}